+++ test/fnq_event_test.dart
@@ -2,6 +2,10 @@ import 'package:flutter_notification_queue/flutter_notification_queue.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 void main() {
+  setUpAll(() {
+    TestWidgetsFlutterBinding.ensureInitialized();
+  });
+
   setUp(() {
     FlutterNotificationQueue.reset();
     FlutterNotificationQueue.configure();